+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validating Webhook for Piraeus StorageClasses.
+
 ### Breaking
 
 - Use DaemonSet to manage Satellite resources instead of bare Pods. This enables better integration with